bot/requirements.txt
python-telegram-bot==21.10
supabase==2.15.3
